Move RichList file to the Identities folder
