+++ packages/frontend/src/components/identities/index.js
@@ -1,0 +1,7 @@
+import RichList from './RichList'
+import IdentitiesList from './IdentitiesList'
+
+export {
+  RichList,
+  IdentitiesList
+}
